Move consts and shared function to own file
